ISSUE-26 Refactor 'commands' and 'subcommands' package
