+++ src/streetrace/commands/subcommands/__init__.py
@@ -4,7 +4,7 @@
 e.g. configure """
 
 from .base_subcommand import BaseSubcommand
-from .configure_subcommand import ConfigureSubcommand
+from .configure import ConfigureSubcommand
 from .parser import SubcommandParser
 from .registry import SubcommandRegistry
