+++ src/streetrace/commands/subcommands/configure/__init__.py
@@ -1,0 +1,5 @@
+"""Configure subcommand package."""
+
+from .subcommand import ConfigureSubcommand
+
+__all__ = ["ConfigureSubcommand"]
