+++ src/streetrace/commands/subcommands/registry.py
